requirements.txt
django-taggit>=0.17.3
django-model-utils>=2.3.1,<3
git+ssh://git@git.thecut.net.au/thecut-authorship@0.10.1
